fix evaluation code

+++ ssim.py
@@ -4,14 +4,14 @@ from math import exp
 import numpy as np
 
 def gaussian(window_size, sigma):
-    gauss = torch.Tensor([exp(-(x - window_size//2)**2/float(2*sigma**2)) for x in range(window_size)])
+    gauss = torch.Tensor([exp(-(x - window_size//2)**2/float(2*sigma**2)) for x in range(window_size)]).double()
     # gauss.requires_grad = True
     return gauss/gauss.sum()
 
 
 def create_window(window_size):
     _1D_window = gaussian(window_size, 1.5).unsqueeze(1)
-    _2D_window = _1D_window.mm(_1D_window.t()).float().unsqueeze(0).unsqueeze(0)
+    _2D_window = _1D_window.mm(_1D_window.t()).double().unsqueeze(0).unsqueeze(0)
     window = _2D_window.expand(1, 1, window_size, window_size).contiguous()
     return window.cuda()
